routes: change stats route to 'uses' route, and change data sent
