+++ app/models/annotation_text.rb
@@ -15,22 +15,25 @@ class AnnotationText < ApplicationRecord
            .gsub(/\r?\n/, '\\n')
   end
 
-  def stats
+  def uses
     # TODO: simplify second join once creator is no longer polymoprhic
-    applications = self.annotations.joins(result: { grouping: :group })
-                       .joins('INNER JOIN users ON annotations.creator_id = users.id')
-                       .order('groups.group_name')
-                       .pluck_to_hash('results.id AS result_id',
-                                      'groupings.assessment_id AS assignment_id',
-                                      'results.submission_id AS submission_id',
-                                      'groups.group_name AS group_name',
-                                      'users.first_name AS first_name',
-                                      'users.last_name AS last_name',
-                                      'users.user_name AS user_name')
-    stats = {
-      num_times_used: self.annotations.count,
-      uses: applications
-    }
-    stats
+    return self.annotations.joins(result: { grouping: :group })
+                           .joins('INNER JOIN users ON annotations.creator_id = users.id')
+                           .order('groups.group_name')
+                           .group('results.id',
+                                  'groupings.assessment_id',
+                                  'results.submission_id',
+                                  'groups.group_name',
+                                  'users.first_name',
+                                  'users.last_name',
+                                  'users.user_name')
+                           .pluck_to_hash('results.id AS result_id',
+                                          'groupings.assessment_id AS assignment_id',
+                                          'results.submission_id AS submission_id',
+                                          'groups.group_name AS group_name',
+                                          'users.first_name AS first_name',
+                                          'users.last_name AS last_name',
+                                          'users.user_name AS user_name',
+                                          'count(*) AS count')
   end
 end
