+++ app/views/annotation_categories/_annotation_text.html.erb
@@ -31,7 +31,7 @@
         <%= annotation_text.last_editor.user_name %>
       <% end %>
     </div>
-    <div id='annotation-text-<%= annotation_text.id %>-stat-panel'></div>
+    <div id='annotation-text-<%= annotation_text.id %>-usage-panel'></div>
   </div>
 
   <div id='annotation_text_<%= annotation_text.id %>_edit' style='display:none;'>
@@ -55,13 +55,15 @@
 </li>
 
 <script>
-  statsPanel = document.getElementById('annotation-text-' + <%= annotation_text.id %> + '-stat-panel');
-  window.annotationStatPanel = window.makeAnnotationStatPanel(
-    statsPanel,
-    {
-      annotation_id: <%= annotation_text.id %>,
-      assignment_id: <%= annotation_text.annotation_category.assessment_id %>,
-      num_used: <%= annotation_text.annotations.count %>
-    }
-  );
+  (() => {
+    let usagePanel = document.getElementById('annotation-text-' + <%= annotation_text.id %> + '-usage-panel');
+    window.AnnotationUsagePanel = window.makeAnnotationUsagePanel(
+      usagePanel,
+      {
+        annotation_id: <%= annotation_text.id %>,
+        assignment_id: <%= annotation_text.annotation_category.assessment_id %>,
+        num_used: <%= annotation_text.annotations.count %>
+      }
+    );
+  })();
 </script>
